// src/app/models/newUser.ts
export class NewUser {

    name: string;
    nricType: number;
    nric_number: any;
    passport_number: string;
    date_of_birth: Date;
    gender: string;
    placeBirth: number;
    otherPlaceOfBirth:string;
    nationality: number;
    ic_address: string;
    address1: string;
    address2: string;
    postalCode: number;
    city: string;
    state: string;
    email_address: string;
    mobile_number: string;
    race: number;
    otherRace:string;
    preference: number;
    emergency_name: string;
    emergency_number: string;
    existingLicence: string;
    user_name: string;
    password: string;
    photo: string;
    nricFront: any;
    nricRear: any;
    passportFrnt: any;
    workPermitFrnt: any;
    existLicFrnt: any;
    existLicRear: any;
    apiToken: any;
    existingLicense: any;
    licenseExpiryDate: Date;
    

}